# Conflicts:
#	ipc/src/main/java/com/sunmi/ipc/cash/CashVideoOverviewActivity.java

+++ app/src/main/java/com/sunmi/assistant/dashboard/card/OverviewDataCard.java
@@ -341,8 +341,8 @@ public class OverviewDataCard extends BaseRefreshCard<OverviewDataCard.Model, Ob
         private String mNum100Million;
         private String mNum10Thousands;
 
-        float sales;
-        float lastSales;
+        double sales;
+        double lastSales;
         int volume;
         int lastVolume;
         int customer;
